Switch from using QA to QC everywhere
Using "QA" everywhere from the beginning was my mistake and
one that has bugged me for a while.  QA in the name QATrack+ is
appropriate because QATrack+ is a process oriented tool for
making sure QC is done on time and capturing QC data.

It's annoying that objects and models will likely need to continue
to use "QA" but this change should at least allow us to use QC in
the UI and docs.

Sorry to all those who are annoyed by this in the future :p