Add different marker size for data points with comments
If a charted test instance (or its parent test list instance) point has
a comment, it will be shown with a slightly larger marker in a blue
colour.  The user can then "lock" the Test Instance tooltip, and hover
over the comment icon to get a popover containing the comment.

Resolves issue #230