Merge Multi Leaf Consulting changes
This merge brings in some of the downstream changes made to QATrack+
for Multi Leaf Consultings hosting service. Notably the reports
system and some of the changes to the notification system.  Further
modifications to the notification system will be merged in later.

More work is required to get reports up and running, periodic/scheduled
tasks with django-q on Windows (not tested at all) etc.