Exclude method lookups from dependency graph
Previously if you had a calc procedure like:
c = Foo()
a = {
    'b': c.d,
}

and another calc procedure like:

d = a['b']

You would get a cyclic dependency because of the ".d" in the first
caclulation (since it would be tokenized as ["b", ".", "d"]). By
excluding anything which is preceded by a "." token we can eliminate
this problem.