Include inactive UTC test lists where required for service events
If a RTSQA object refers to a test list instance from an
inactive UTC, it will now be visible in the Return To Service QC
section.  Previously it would not be displayed since the UTC
queryset had an active=True filter.  I think active=True should
remain the default queryset, so have opted to only include
inactive UTCS when the RTSQA has already been performed.

Resolves issue #427