change order of sql commands
In case of a postgresql upgrade, the previous database might have been dropped. somehow if a new db is created and the backup inserted *before* rights are granted to user qatrack, privileges are not given to tables...
but if you create the db, create the user, grant privileges, and only then insert the backup in the new db, then it works fine.