Attempt to fix issue with datepicker sometimes not being initiated properly for service events
Make base.js a dependency of select2 to ensure overrideSelect2Keys  is
available when its called on select2 elements.  There were very rare
occassions when $service_status.select2({...}).overrideSelect2Keys();
was giving an "overrideSelect2Keys is not a function" error. This
patch should (?) resolve this.

Also, if the user clicks on the input before the javascript is finished
loading the datepicker would not be shown.  This patch addresses that
by checking if the input is focused at the time the datepicker is
initialized and then opening it if it is.